Storybook page created. Component renamed from ColorPalettes to SwatchPaletteWidget so that css classes do not clash with [ibm-col] classes Test mdx page named palettes created in color section 2 components added to src/components, one for each widget, since mdx wont pass an array from front-matter without some extensive work in global space. Alternative would be to parse a string into one or two arrays inside the component and pass a string into each component in the mdx pages
