+++ packages/addons-website/scss/components/_swatch-palette-widget.scss
@@ -1,0 +1,89 @@
+@import '../prefix';
+
+.#{$prefix}--swatch-palettes {
+  display: flex;
+  flex-direction: column;
+  .#{$prefix}--swatch-palette {
+    float: left;
+    width: 100%;
+    margin-bottom: 2rem;
+    .#{$prefix}--swatch {
+      padding: 1rem;
+      font-family: IBM Plex Mono,Menlo,DejaVu Sans Mono,Bitstream Vera Sans Mono,Courier,monospace;
+      .#{$prefix}--swatch-name {
+        text-transform: capitalize;
+      }
+      .#{$prefix}--swatch-value {
+        float: right;
+      }
+    }
+    
+    @include breakpoint('md') {
+      width: calc(50% - 1px);
+      margin-right: 1px; 
+      &.last-child {
+        width: 50%;
+        margin-right: 0; 
+      }
+    }
+    @include breakpoint('lg') {
+      width: calc(33.33% - 1px);
+      margin-right: 1px; 
+      &.last-child {
+        width: 33.33%;
+        margin-right: 0; 
+      }
+    }
+  }
+}
+
+.#{$prefix}--swatch-palettes-container {
+  position: relative;
+  width: 100%;
+  margin-bottom: 2rem;
+
+  .sticky-container {
+    position: sticky;
+    z-index: 8888;
+    margin: 0 -1rem;
+    padding: 0 1rem;
+    background: $ui-01;
+    top: 0;
+    @include breakpoint('lg') {
+      padding-top: 1rem;
+    }
+  }
+}
+
+.#{$prefix}--swatch-palettes__format-switcher {
+  position: relative;
+  width: 100%;
+  height: rem(49px);
+  padding-bottom: 1px;
+
+  .#{$prefix}--content-switcher-btn {
+    flex: 1;
+    width: calc(25% - 1px);
+    margin-right: 1px;
+    height: rem(48px);
+    padding: .5rem 1rem;
+    outline: none;
+    transition: background-color 100ms cubic-bezier(0.2, 0.2, 0.38, 0.9);
+
+    &:not(:last-of-type)::after {
+      height: 100%;
+      right: -1px;
+      top: 0;
+    }
+
+    &:not(.#{$prefix}--content-switcher--selected) {
+      background-color: $ui-02;
+      color: $text-01;    
+      &:hover {
+        background-color: $hover-ui;
+      }
+    }
+  }
+}
+
+
