+++ packages/addons-website/scss/styles.scss
@@ -6,3 +6,4 @@
 @import 'components/website-header-nav';
 @import 'components/website-side-nav';
 @import 'components/website-back-to-top-btn';
+@import 'components/swatch-palette-widget';
