+++ packages/addons-website/src/components/SwatchPaletteWidget/index.js
@@ -1,0 +1,2 @@
+import SwatchPaletteWidget from './SwatchPaletteWidget';
+export default SwatchPaletteWidget;
