--- src/components/ColorPalette/index.js
@@ -1,3 +1,0 @@
-import ColorPalette from './ColorPalette';
-
-export default ColorPalette;
