--- src/components/ColorPaletteErrors/index.js
@@ -1,3 +1,0 @@
-import ColorPaletteErrors from './ColorPaletteErrors';
-
-export default ColorPaletteErrors;
