--- src/components/ColorPalettes/color-palettes.scss
@@ -1,107 +1,0 @@
-// .color-palettes {
-//   display: flex;
-//   flex-direction: column;
-//   .color-palette {
-//     float: left;
-//     width: calc(33.33% - 1px);
-//     margin-right: 1px; 
-//     margin-bottom: 2rem;
-//     .color-swatch {
-//       padding: 1rem;
-//       font-family: IBM Plex Mono,Menlo,DejaVu Sans Mono,Bitstream Vera Sans Mono,Courier,monospace;
-//       .color-name {
-//         text-transform: capitalize;
-//       }
-//       .color-value {
-//         float: right;
-//       }
-//     }
-//     @media all and (max-width: 1055px) {
-//       width: calc(50% - 1px);
-//       &.last-child {
-//         width: 50%;
-//       }
-//     }
-//     @media all and (max-width: 671px) {
-//       width: 100%;
-//       margin: 0; 
-//     }
-//   }
-// }
-
-.color-palettes__format-switcher {
-  position: relative;
-  width: 100%;
-  height: rem(49px);
-  padding-bottom: 1px;
-  // background: $ui-02;
-  background: #f3f3f3;
-}
-
-// .color-palettes__format-switcher::before {
-//   content: '';
-//   display: block;
-//   position: absolute;
-//   top: 0;
-//   left: 0;
-//   width: 110%;
-//   height: 64px;
-//   background: $ui-01;
-// }
-
-
-
-.color-palettes__format-switcher
-  .bx--content-switcher-btn:not(.bx--content-switcher--selected) {
-  // background-color: $ui-02;
-  background-color: #ffffff;
-  color: $text-01;
-
-  &:hover {
-    background-color: #e5e5e5; // $hover-ui
-  }
-}
-
-.color-palettes__format-switcher .bx--content-switcher-btn {
-  flex: 1;
-  width: calc(25% - 1px);
-  margin-right: 1px;
-  height: rem(48px);
-  padding: .5rem 1rem;
-  outline: none;
-  transition: background-color 100ms cubic-bezier(0.2, 0.2, 0.38, 0.9);
-}
-
-.color-palettes__format-switcher
-  .bx--content-switcher-btn:not(:last-of-type)::after {
-  height: 100%;
-  right: -1px;
-  top: 0;
-}
-.color-palettes__format-switcher--sticky {
-  // background: $ui-02;
-  position: fixed;
-  top: 88px;
-  z-index: 9999;
-  width: 100%;
-  .bx--content-switcher-btn {
-    flex: none;
-    width: calc(25% - 1px);
-  }
-
-  @include breakpoint('md') {
-    width: calc(100% - 2rem);
-    .bx--content-switcher-btn {
-      width: calc(12.5% - 1px);
-    }
-  }
-
-  @include breakpoint('lg') {
-    width: calc(75% - 1.5rem);
-    max-width: 1164px; 
-    .bx--content-switcher-btn {
-      width: calc(8.33% - 1px);
-    }
-  }
-  // box-shadow: 0px 2px 6px 0px rgba(#000000, 0.2);
-}
