--- src/components/ColorPalettes/index.js
@@ -1,3 +1,0 @@
-import ColorPalettes from './ColorPalettes';
-
-export default ColorPalettes;
