+++ src/components/SwatchPaletteWidget1/index.js
@@ -1,0 +1,3 @@
+import SwatchPaletteWidget1 from './SwatchPaletteWidget1';
+
+export default SwatchPaletteWidget1;
