+++ src/components/SwatchPaletteWidget2/index.js
@@ -1,0 +1,3 @@
+import SwatchPaletteWidget2 from './SwatchPaletteWidget2';
+
+export default SwatchPaletteWidget2;
