+++ src/content/guidelines/color/overview.mdx
@@ -1,6 +1,6 @@
 ---
 title: Color
-tabs: ['Overview', 'Usage', 'Palettes']
+tabs: ['Overview', 'Usage']
 ---
 
 ### Color is a foundational element of Carbon. Careful application of color drives consistency, engagement, and focus for all user interfaces.
