+++ src/content/guidelines/color/palettes.mdx
@@ -3,18 +3,19 @@ title: Color
 tabs: ['Overview', 'Usage', 'Palettes']
 ---
 
-import ColorPalettes from '../../../../src/components/ColorPalettes';
+import SwatchPaletteWidget1 from '../../../../src/components/SwatchPaletteWidget1'
+import SwatchPaletteWidget2 from '../../../../src/components/SwatchPaletteWidget2'
 
 ## Specifications
 
 Each of the 10 color families have been divided into 10 swatches ranging from light to dark. RGB and HEX values are provided for digital appliations along with Pantone and CMYK values for print.
 
-<ColorPalettes />
 
-import ColorPaletteErrors from '../../../../src/components/ColorPaletteErrors';
+<SwatchPaletteWidget1 />
 
 ## Specifications
 
 The Neutral Gray family is dominant in our UI, making use of subtle shifts in value to help organize content into distinct zones. The Core Blue family has been designated as the primary action color across all products and experiences, ensuring our color aesthetic is a part of every interaction. Additional colors are used sparingly and purposefully.
 
-<ColorPaletteErrors />
+
+<SwatchPaletteWidget2 />
