+++ src/content/guidelines/color/usage.mdx
@@ -1,7 +1,7 @@
 ---
 label: Color is a foundational element of Carbon. Careful application of color drives consistency, engagement, and focus for all user interfaces.
 title: Color
-tabs: ['Overview', 'Usage', 'Palettes']
+tabs: ['Overview', 'Usage']
 ---
 
 import ColorTokenTable from '../../../../src/components/ColorTokenTable';
