+++ src/styles/index.scss
@@ -71,8 +71,6 @@ $font-family: 'IBM Plex Sans', 'Helvetica Neue', Arial, sans-serif;
 @import '../components/FeatureTile/feature-tile.scss';
 @import '../components/WebsiteTabs/website-tabs.scss';
 @import '../components/ColorTokenTable/color-token-table.scss';
-@import '../components/ColorPalette/color-palette.scss';
-@import '../components/ColorPalettes/color-palettes.scss';
 @import '../components/SimpleColumns/simple-columns.scss';
 @import '../components/ImageComponent/image-component.scss';
